### 2024-01-22 #6242 @Kevin-Mattheus-Moerman
@JaGeo thanks for this submission. I am the AEiC on this track and here to help with initial steps.

To fellow editors. There is a [prior publication](https://chemistry-europe.onlinelibrary.wiley.com/doi/10.1002/cplu.202200123). For clarity I reproduce the notes to the editor that the authors provided on this:

> This package provides valuable tools to summarize the LOBSTER software outputs in JSON, making them easy to interpret and process. The initial version of our package with limited capabilities played a crucial role in generating results of our prior work titled "Automated Bonding Analysis with Crystal Orbital Hamilton Populations (DOI: 10.1002/cplu.202200123)" published in the ChemPlusChem journal. Since then, we have extended the package's core functionalities significantly. This includes support for automated analysis of the Crystal Orbital Bond index and the Crystal Orbital Overlap Populations and updates to automatic analysis outputs such as estimating bonding-antibonding contributions and orbital-wise analysis. We have also added customizable interactive plots, more plotting options (ICOHPs vs bond lengths, density of states plots), support for other codes besides VASP, and more robust testing and documentation. Additionally, two new submodules that enable feature engineering for machine learning studies are added.
