### 2024-01-26 #6242 @RMeli
@berquist 👋 Would you be interested in reviewing this submission for JOSS?
