### 2024-01-26 #6242 @RMeli
@srmnitc 👋 Would you be interested in reviewing this submission for JOSS?
