### 2024-01-26 #6242 @srmnitc
> @srmnitc 👋 Would you be interested in reviewing this submission for JOSS?

Sure, seems interesting. Thanks for the invite!
