### 2024-01-26 #6286 @editorialbot
[REVIEW]: LobsterPy: A package to automatically analyze LOBSTER runs
**Submitting author:** @JaGeo (<a href="http://orcid.org/0000-0001-8907-0336">Janine George</a>)
**Repository:** https://github.com/JaGeo/LobsterPy
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.8
