### 2024-02-07 JaGeo/LobsterPy#221 @srmnitc
pip installation fails
**Describe the bug**
This is part of the review over at https://github.com/openjournals/joss-reviews/issues/6286. I tried to install using pip in a clean conda environment, and the installation fails with the attached message

**To Reproduce**
Steps to reproduce the behavior:
1. create new conda env
2. install pip
3. run `pip install lobsterpy`

**Expected behavior**
Should install the package

**Screenshots**
```

  Building wheel for spglib (pyproject.toml) ... error
  error: subprocess-exited-with-error

  × Building wheel for spglib (pyproject.toml) did not run successfully.
  │ exit code: 1
  ╰─> [37 lines of output]
      /tmp/pip-build-env-esh8lx6s/normal/lib/python3.12/site-packages/setuptools_scm/git.py:308: UserWarning: git archive did not support describe output
        warnings.warn("git archive did not support describe output")
      /tmp/pip-build-env-esh8lx6s/normal/lib/python3.12/site-packages/setuptools_scm/git.py:327: UserWarning: unprocessed git archival found (no export subst applied)
        warnings.warn("unprocessed git archival found (no export subst applied)")
      *** scikit-build-core 0.8.0 using CMake 3.16.3 (wheel)
      *** Configuring CMake...
      2024-02-07 09:51:59,847 - scikit_build_core - WARNING - libdir/ldlibrary: /home/menon/miniconda3/envs/lobsterpy/lib/libpython3.12.a is not a real file!
      2024-02-07 09:51:59,847 - scikit_build_core - WARNING - Can't find a Python library, got libdir=/home/menon/miniconda3/envs/lobsterpy/lib, ldlibrary=libpython3.12.a, multiarch=x86_64-linux-gnu, masd=None
      loading initial cache file /tmp/tmpgrjd0cxd/build/CMakeInit.txt
      CMake Error: Error processing file: COMMAND_ERROR_IS_FATAL
      CMake Error at cmake/DynamicVersion.cmake:270 (file):
        file failed to open for reading (No such file or directory):

          /tmp/tmpgrjd0cxd/build/tmp/.DynamicVersion.json
      Call Stack (most recent call first):
        CMakeLists.txt:14 (dynamic_version)

      CMake Warning at CMakeLists.txt:18 (project):
        VERSION keyword not followed by a value or was followed by a value that
        expanded to nothing.

      -- The C compiler identification is GNU 9.3.0
      -- Check for working C compiler: /usr/bin/cc
      -- Check for working C compiler: /usr/bin/cc -- works
      -- Detecting C compiler ABI info
      -- Detecting C compiler ABI info - done
      -- Detecting C compile features
      -- Detecting C compile features - done
      -- Spglib:: Build type: Release
      -- Spglib:: Version
      -- Found Python3: /home/menon/miniconda3/envs/lobsterpy/bin/python3.12 (found version "3.12.1") found components: Interpreter Development.Module NumPy
      -- Configuring incomplete, errors occurred!
      See also "/tmp/tmpgrjd0cxd/build/CMakeFiles/CMakeOutput.log".

      *** CMake configuration failed
      [end of output]

  ERROR: Failed building wheel for spglib
  Building wheel for future (setup.py) ... done
  Created wheel for future: filename=future-0.18.3-py3-none-any.whl size=492024 sha256=5d5097157e5f196a515ad1e810b8ad8139b38c60699d49591587c34fa5e1a2ca
  Stored in directory: /home/menon/.cache/pip/wheels/f1/54/27/4e5cb4bd05116c097877e85a206a2e08d301aaa9b7af719909
Successfully built pymatgen typing future
Failed to build spglib
ERROR: Could not build wheels for spglib, which is required to install pyproject.toml-based projects
```

 - WSL with Ubuntu 20
