### 2024-02-07 JaGeo/LobsterPy#221 @naik-aakash
Hi @srmnitc , thank you for reporting this.

Can you please check if the issue still persist if using any python version <3.12 ? For example 3.11 or so?

Seems the problem is not specific to lobsterpy but originates from pymatgen dependencies. And using python 3.12, I also encounter the same issue, but with python 3.11, it does not occur. (Seems many packages are not yet fully adapted to support latest python version)
