### 2024-02-07 JaGeo/LobsterPy#221 @JaGeo
@naik-aakash can we be somehow clearer in which Python versions we support?
