### 2024-02-07 JaGeo/LobsterPy#221 @naik-aakash
> @naik-aakash can we be somehow clearer in which Python versions we support?

We already have this mentioned which python versions work in our pyproject.toml https://github.com/JaGeo/LobsterPy/blob/8e777b85a71f8e128ae141449db456a25ec94f3f/pyproject.toml#L20 and also visible on pypi https://pypi.org/project/lobsterpy/.

I can think of adding only hard coded limit here to avoid such issues in future> https://github.com/JaGeo/LobsterPy/blob/8e777b85a71f8e128ae141449db456a25ec94f3f/pyproject.toml#L29
