### 2024-02-07 JaGeo/LobsterPy#221 @srmnitc
Thanks! I was indeed using python 3.12. The installation works fine for 3.11. I can see now that the versions are also clearly defined, however, I just did a pip install, and it did not complain about the python version, but rather crashed. Maybe it is better to specify this. That being said, feel free to close this issue. Thanks for the quick resolution.
