### 2024-02-07 JaGeo/LobsterPy#221 @JaGeo
We might want to set a maximum Python version, though. Thanks!

@naik-aakash Could you check if works?
