+++ pyproject.toml
@@ -26,7 +26,7 @@ classifiers = [
     "Intended Audience :: Science/Research",
     "Operating System :: OS Independent",
 ]
-requires-python = ">=3.9"
+requires-python = ">=3.9,<3.12"
 dependencies = [
      "pymatgen>=2023.9.10",
      "numpy",
