### 2024-02-07 4efd810
disable potcar parse when reading vasprun
