+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
+## v0.3.7
+* Set python version stricter limit and accommodate recent pymatgen changes by @naik-aakash in https://github.com/JaGeo/LobsterPy/pull/222
+
 ## v0.3.6
 * JOSS paper by @naik-aakash in https://github.com/JaGeo/LobsterPy/pull/203
-
 
 ## v0.3.5
 * Fix header by @naik-aakash in https://github.com/JaGeo/LobsterPy/pull/189
