### 2024-02-08 JaGeo/LobsterPy#223 @srmnitc
Information on how to run the tests
This is part of the review over at https://github.com/openjournals/joss-reviews/issues/6286. The project seems to have automated tests, but I could not find instructions on how a user can run the tests. This would allow me to check off the 'Automated Tests' in the review.

**Describe the solution you'd like**
Add a documentation on how to run tests in both README and documentation page

None

**Additional context**
None
