### 2024-02-08 JaGeo/LobsterPy#223 @JaGeo
Thank you, @srmnitc . Could you check if this description works: https://jageo.github.io/LobsterPy/dev/dev_installation.html ?

If not, please let us know and we are happy to help.
