### 2024-02-08 JaGeo/LobsterPy#223 @JaGeo
Meanwhile, @naik-aakash and I will be thinking about how to make this documentation part more discoverable.
