### 2024-02-08 JaGeo/LobsterPy#223 @srmnitc
Perfect @JaGeo! That works, I was able to run the tests. Once again, a link to the dev docs would be nice, but everything works very well. Feel free to close the issue.
