### 2024-02-08 JaGeo/LobsterPy#223 @JaGeo
Could you please download the latest main branch, upgrade the pymatgen version and check if the error persists? I assume it is a related to a very recent upgrade in pymatgen.
