### 2024-02-08 JaGeo/LobsterPy#223 @naik-aakash
Hi @srmnitc , this is result of recent change in pymatgen. We relased a new version of LobsterPy yesterday that address this issue. Try upgrading the lobsterpy.
