### 2024-02-08 JaGeo/LobsterPy#223 @srmnitc
> Could you please download the latest main branch, upgrade the pymatgen version (`pip install --upgrade pymatgen`) and check if the error persists? I assume it is a related to a very recent upgrade in pymatgen. @naik-aakash fixed that error yesterday here on the main branch.

This solution worked, thanks.
