### 2024-02-08 4ec6c7f
add devloper installation/contributing guide link to readme. Reorder Developer Installation, Contributing guide in doc index
