### 2024-02-09 JaGeo/LobsterPy#225 @srmnitc
Suggestions for JOSS paper
Here are some suggestions for the JOSS paper:

- citation for LOBSTER
- are there any tools similar to lobsterpy in the community?
