### 2024-02-09 #6286 @srmnitc
@RMeli I have now finished my review. Lobsterpy is a well-designed, and maintained package which I believe would be useful for the community. It has already been employed in a number of studies. The code and documentation are in very good shape, and it was really easy to get started with the tool. I had only some minor issues, which @JaGeo and team fixed very quickly. There are only minor corrections to the paper remaining. Overall, I recommend the publication of this package. Thanks @RMeli for inviting me to review, and @JaGeo for this nice software.
