### 2024-02-09 #6286 @RMeli
Thank you @srmnitc for all the hard work!

> I had only some minor issues, which @JaGeo and team fixed very quickly.

I've noticed it, thanks @JaGeo for the very prompt replies to the reviewer queries and comments.

> There are only minor corrections to the paper remaining.

Did you suggest these minor corrections anywhere? I only see https://github.com/JaGeo/LobsterPy/issues/225, are you referring just to those two suggestions?
