### 2024-02-09 009c081
address review comments > LOBSTER citation, similar packages
