+++ paper/paper.bib
@@ -1,7 +1,82 @@
+@article{deringer2011crystal,
+  title={Crystal orbital Hamilton population (COHP) analysis as projected from plane-wave basis sets},
+  author={Deringer, Volker L and Tchougr{\'e}eff, Andrei L and Dronskowski, Richard},
+  journal={The journal of physical chemistry A},
+  volume={115},
+  number={21},
+  pages={5461--5466},
+  year={2011},
+  publisher={ACS Publications}
+  doi={10.1021/jp202489s},
+}
+
+@article{maintz2013analytic,
+  title={Analytic projection from plane-wave and PAW wavefunctions and application to chemical-bonding analysis in solids},
+  author={Maintz, Stefan and Deringer, Volker L and Tchougr{\'e}eff, Andrei L and Dronskowski, Richard},
+  journal={Journal of computational chemistry},
+  volume={34},
+  number={29},
+  pages={2557--2567},
+  year={2013},
+  publisher={Wiley Online Library},
+  doi={10.1002/jcc.23424},
+}
+
+@misc{maintz2016lobster,
+  title={LOBSTER: A tool to extract chemical bonding from plane-wave based DFT},
+  author={Maintz, Stefan and Deringer, Volker L and Tchougr{\'e}eff, Andrei L and Dronskowski, Richard},
+  journal = {Journal of Computational Chemistry},
+  volume = {37},
+  number = {11},
+  pages = {1030--1035},
+  year={2016},
+  publisher={Wiley Online Library},
+  doi={10.1002/jcc.24300},
+}
+
+@article{nelson2020lobster,
+  title={LOBSTER: Local orbital projections, atomic charges, and chemical-bonding analysis from projector-augmented-wave-based density-functional theory},
+  author={Nelson, Ryky and Ertural, Christina and George, Janine and Deringer, Volker L and Hautier, Geoffroy and Dronskowski, Richard},
+  journal={Journal of Computational Chemistry},
+  volume={41},
+  number={21},
+  pages={1931--1940},
+  year={2020},
+  publisher={Wiley Online Library}
+}
+
 @misc{atomate2,
 title = {Atomate2},
 year = {2023},
 url = {https://github.com/materialsproject/atomate2}
+}
+
+@misc{abipy,
+title = {abipy},
+year = {2021},
+url = {https://github.com/abinit/abipy}
+}
+
+@misc{EckwxDragon,
+	author = {Eck, Bernhard},
+	title = {wxDragon, Version 2.2.3},
+	publisher = {RWTH Aachen University},
+	address = {Aachen, Germany},
+	year = {1994–2020},
+	url = {https://schmeling.ac.rwth-aachen.de/cohp/index.php?menuID=6}
+}
+
+
+@article{Ganose2018,
+  title = {sumo: Command-line tools for plotting and analysis of periodic *ab initio* calculations},
+  author = {Alex M. Ganose and Adam J. Jackson and David O. Scanlon},
+  journal = {Journal of Open Source Software},
+  volume = {3},
+  number = {28},
+  pages = {717},
+  year = {2018},
+  publisher = {The Open Journal},
+  doi = {10.21105/joss.00717},
 }
 
 @article{ong2013python,
