+++ paper/paper.bib
@@ -57,7 +57,7 @@ year = {2021},
 url = {https://github.com/abinit/abipy}
 }
 
-@misc{EckwxDragon,
+@misc{wxdragon,
 	author = {Eck, Bernhard},
 	title = {wxDragon, Version 2.2.3},
 	publisher = {RWTH Aachen University},
