### 2024-02-09 2f11ef5
Update paper/paper.md
Co-authored-by: J. George <JaGeo@users.noreply.github.com>
