### 2024-02-09 77a4cf2
Update paper/paper.md
Co-authored-by: J. George <JaGeo@users.noreply.github.com>
