+++ paper/paper.bib
@@ -59,7 +59,7 @@ url = {https://github.com/abinit/abipy}
 
 @misc{wxdragon,
 	author = {Eck, Bernhard},
-	title = {wxDragon, Version 2.2.3},
+	title = {wxDragon, Version 2.2.2},
 	publisher = {RWTH Aachen University},
 	address = {Aachen, Germany},
 	year = {1994–2020},
