### 2024-02-09 JaGeo/LobsterPy#225 @JaGeo
@srmnitc , @naik-aakash and I have addressed your comments in #226 . Are you fine with the changes?
