### 2024-02-09 JaGeo/LobsterPy#225 @JaGeo
Great, @srmnitc. We will merge then and close this issue (cc @RMeli )
