### 2024-02-09 #6286 @JaGeo
We (@naik-aakash and I) have addressed the comments in https://github.com/JaGeo/LobsterPy/pull/226, @RMeli .
