+++ paper/paper.bib
@@ -6,7 +6,7 @@
   number={21},
   pages={5461--5466},
   year={2011},
-  publisher={ACS Publications}
+  publisher={ACS Publications},
   doi={10.1021/jp202489s},
 }
 
