### 2024-02-21 f7df19b
add note about tutorial > juypter notebook
