+++ docs/tutorial/index.rst
@@ -19,6 +19,8 @@ This tutorial will demonstrate how to use the LobsterPy package using example co
 4. Use featurizers to extract the LOBSTER bonding analysis data as features for the ML studies
 5. Get automatic analysis results and plots using command line utilities
 
+**Note**: The following tutorial is jupyter Notebook, you can download it and execute locally. Please ignore the code blocks which have `remove-cell` tags. You can see these cell tags using `View -> Cell Toolbar -> Tags` (These code blocks are hidden in rendered docs to keep it consistent with example files paths you will use.)
+
 .. toctree::
    :maxdepth: 2
 
