### 2024-02-21 2d72b92
adapt LobsterEnv usage to breaking change
