### 2024-02-22 bb6d609
add analyse pymatgen objs fixtures
