### 2024-02-22 b17b0ae
adapt analyze to work with pymatgen objects
