+++ lobsterpy/cohp/analyze.py
@@ -13,7 +13,15 @@ from pymatgen.analysis.bond_valence import BVAnalyzer
 from pymatgen.core.structure import Structure
 from pymatgen.electronic_structure.cohp import CompleteCohp
 from pymatgen.electronic_structure.core import Spin
-from pymatgen.io.lobster import Bandoverlaps, Charge, Doscar, Lobsterin, Lobsterout
+from pymatgen.io.lobster import (
+    Bandoverlaps,
+    Charge,
+    Doscar,
+    Icohplist,
+    Lobsterin,
+    Lobsterout,
+    MadelungEnergies,
+)
 from pymatgen.io.lobster.lobsterenv import LobsterNeighbors
 from pymatgen.io.vasp.outputs import Vasprun
 from pymatgen.symmetry.analyzer import SpacegroupAnalyzer
@@ -27,11 +35,15 @@ class Analysis:
     :param are_coops: bool indicating if file contains COOP/ICOOP data
     :param cutoff_icohp: Cutoff in percentage for evaluating neighbors based on ICOHP values.
         cutoff_icohp*max_icohp limits the number of considered neighbours for evaluating environments.
-    :param path_to_charge: path to `CHARGE.lobster`.
     :param path_to_cohpcar: path to `COHPCAR.lobster` or `COBICAR.lobster` or `COOPCAR.lobster` .
+    :param path_to_charge: path to `CHARGE.lobster`.
     :param path_to_icohplist: path to `ICOHPLIST.lobster` or `ICOBILIST.lobster` or `ICOOPLIST.lobster`.
     :param path_to_poscar: path to structure (e.g., `POSCAR` or `POSCAR.lobster`)
     :param path_to_madelung: path to `MadelungEnergies.lobster`.
+    :param charge_obj: pymatgen lobster.io.charge object
+    :param completecohp_obj: pymatgen.electronic_structure.cohp.CompleteCohp object
+    :param icohplist_obj: pymatgen lobster.io.Icohplist object
+    :param madelung_obj: pymatgen lobster.io.MadelungEnergies object
     :param noise_cutoff: Sets the lower limit tolerance for ICOHPs or ICOOPs or ICOBIs considered
         in analysis.
     :param orbital_cutoff: Sets the minimum percentage for the orbital contribution considered to be
@@ -70,11 +82,15 @@ class Analysis:
 
     def __init__(
         self,
-        path_to_poscar: str | Path,
-        path_to_icohplist: str | Path,
-        path_to_cohpcar: str | Path,
+        path_to_poscar: str | Path | None,
+        path_to_icohplist: str | Path | None,
+        path_to_cohpcar: str | Path | None,
         path_to_charge: str | Path | None = None,
         path_to_madelung: str | Path | None = None,
+        icohplist_obj: Icohplist | None = None,
+        completecohp_obj: CompleteCohp | None = None,
+        charge_obj: Charge | None = None,
+        madelung_obj: MadelungEnergies | None = None,
         are_cobis: bool = False,
         are_coops: bool = False,
         cutoff_icohp: float = 0.1,
@@ -93,11 +109,15 @@ class Analysis:
         :param are_coops: bool indicating if file contains COOP/ICOOP data
         :param cutoff_icohp: Cutoff in percentage for evaluating neighbors based on ICOHP values.
             cutoff_icohp*max_icohp limits the number of considered neighbours for evaluating environments.
-        :param path_to_charge: path to `CHARGE.lobster`.
         :param path_to_cohpcar: path to `COHPCAR.lobster` or `COBICAR.lobster` or `COOPCAR.lobster` .
+        :param path_to_charge: path to `CHARGE.lobster`.
[…]
