+++ lobsterpy/cohp/describe.py
@@ -256,7 +256,11 @@ class Description:
                     if orb_bonds:
                         self.text.append(orb_bonds)
 
-        if "madelung_energy" in self.analysis_object.condensed_bonding_analysis:
+        if (
+            "madelung_energy" in self.analysis_object.condensed_bonding_analysis
+            and self.analysis_object.condensed_bonding_analysis["madelung_energy"]
+            is not None
+        ):
             self.text.append(
                 "The Madelung energy of this crystal structure per unit cell is: "
                 + str(
