### 2024-02-22 1600a31
add tests for pymatgen objs
