### 2024-02-22 fe3ebb2
add sanity check
