### 2024-02-23 7e879d8
Revert "adapt LobsterEnv usage to breaking change"
This reverts commit 2d72b9204cc6cdae033d5e14fb9f62e0f909ed5a.
