### 2024-02-23 #6286 @berquist
# Review

I've finished reviewing the code, documentation, and the paper, and am happy to recommend LobsterPy for submission to JOSS.  My background is in molecular electronic structure, not materials or periodic systems, but there is clearly a substantial amount of work done here.  I also get the impression that a lot of this work is hiding in atomate2.

The sections that follow are not required for acceptance, but I think will make the code friendlier for both users and developers, even if those developers are not external contributors.

## Documentation

Reading https://jageo.github.io/LobsterPy/reference/cli_subcommands/createinputs.html there is the statement

> It works only with PBE POTCARs.

that doesn't appear when running `lobsterpy create-inputs --help`.  More importantly, no justification is given for the statement, but I think I found it in https://doi.org/10.1002/cplu.202200123

> For the time being, two basis sets (Koga and pbeVASPfit2015) can be used to cover the whole periodic table, although only pbeVASPfit2015 includes additional orbitals important to the solid state (e. g., 2p in metallic Be[21]).

I would at least reproduce this comment in the help string, and ideally put more of the reasoning somewhere in the documentation, perhaps under a general "limitations" section.

For creating inputs, it would be good to have a simple example for a common case: it wasn't clear to me if the test data used by `test_cli.py`  is there only for edge cases or for the common case as well.

## Developer experience

Some suggestions on the developer experience:

- The unit tests take a while to run (8 minutes in serial, 1.5 minutes with `-n 8 --dist worksteal` on a 16 core machine).  The provided suggestion to use `xdist` is a good one.  If you find that developers complain still, consider using pytest markers to group tests by time.

- The top-level `requirements.txt` doesn't seem to be used anywhere.  Since it is already covered properly by the `pyproject.toml`, consider removing it.

- Consider removing `black` from the pre-commit config and using `- id: ruff-format` since the config is already in `pyproject.toml`.  Similarly, the top-level `pylintrc` is confusing since it isn't part of pre-commit.  Consolidating around tools and cleaning up their configurations will make your lives easier even if it doesn't affect users.

- I strongly recommend adding https://pre-commit.ci/ to the GitHub workflow so that all the pre-commit rules can be enforced for each PR even if someone forgets to install the hook locally.  (I see configuration for it exists, so someone has considered this before, but it's not turned on.)

I am happy to open PRs for any of these if you wish, though adding the CI service is purely an admin action.
