### 2024-02-23 #6286 @RMeli
Thank you @berquist for the review and recommendation!

> The sections that follow are not required for acceptance, but I think will make the code friendlier for both users and developers

@JaGeo (and team), would you be able to look into the additional suggestions and apply them where you see appropriate?
