### 2024-02-23 fd370fb
fix typo in CLI help
