+++ requirements.txt
@@ -1,1 +1,0 @@
-pymatgen>=2023.9.10
