+++ .pre-commit-config.yaml
@@ -9,8 +9,8 @@ ci:
 default_language_version:
   python: python3
 repos:
-- repo: https://github.com/charliermarsh/ruff-pre-commit
-  rev: v0.1.1
+- repo: https://github.com/astral-sh/ruff-pre-commit
+  rev: v0.2.2
   hooks:
   - id: ruff
     args: [--fix]
@@ -23,7 +23,7 @@ repos:
   - id: end-of-file-fixer
   - id: trailing-whitespace
 - repo: https://github.com/psf/black
-  rev: 23.10.0
+  rev: 24.2.0
   hooks:
   - id: black
 - repo: https://github.com/asottile/blacken-docs
@@ -33,7 +33,7 @@ repos:
     additional_dependencies: [black]
     exclude: README.md
 - repo: https://github.com/pycqa/flake8
-  rev: 6.1.0
+  rev: 7.0.0
   hooks:
   - id: flake8
     entry: pflake8
@@ -53,7 +53,7 @@ repos:
   - id: rst-directive-colons
   - id: rst-inline-touching-normal
 - repo: https://github.com/pre-commit/mirrors-mypy
-  rev: v1.6.1
+  rev: v1.8.0
   hooks:
   - id: mypy
     files: lobsterpy/.*
