### 2024-02-23 d84437c
pre-commit: remove black in favor of ruff-format
