+++ .pre-commit-config.yaml
@@ -14,6 +14,7 @@ repos:
   hooks:
   - id: ruff
     args: [--fix]
+  - id: ruff-format
 - repo: https://github.com/pre-commit/pre-commit-hooks
   rev: v4.5.0
   hooks:
@@ -22,10 +23,6 @@ repos:
     args: [--remove]
   - id: end-of-file-fixer
   - id: trailing-whitespace
-- repo: https://github.com/psf/black
-  rev: 24.2.0
-  hooks:
-  - id: black
 - repo: https://github.com/asottile/blacken-docs
   rev: 1.16.0
   hooks:
