+++ lobsterpy/featurize/batch.py
@@ -171,10 +171,7 @@ class BatchSummaryFeaturizer:
             df_cohp = coxx.get_summarized_coxx_df()
             del coxx
         else:
-            raise Exception(
-                "COHPCAR.lobster or POSCAR or ICOHPLIST.lobster file "
-                f"not found in {dir_name.name}"
-            )
+            raise Exception(f"COHPCAR.lobster or POSCAR or ICOHPLIST.lobster file not found in {dir_name.name}")
 
         if self.include_cobi_data:
             req_files = {
@@ -206,10 +203,7 @@ class BatchSummaryFeaturizer:
                 del coxx
 
             else:
-                raise Exception(
-                    "COBICAR.lobster or ICOBILIST.lobster file "
-                    f"not found in {dir_name.name}"
-                )
+                raise Exception(f"COBICAR.lobster or ICOBILIST.lobster file not found in {dir_name.name}")
 
         if self.include_coop_data:
             req_files = {
@@ -241,10 +235,7 @@ class BatchSummaryFeaturizer:
                 del coxx
 
             else:
-                raise Exception(
-                    "COOPCAR.lobster or ICOOPLIST.lobster file "
-                    f"not found in {dir_name.name}"
-                )
+                raise Exception(f"COOPCAR.lobster or ICOOPLIST.lobster file not found in {dir_name.name}")
 
         if self.include_cobi_data and self.include_coop_data:
             df = pd.concat([df_cohp, df_cobi, df_coop], axis=1)
@@ -344,9 +335,7 @@ class BatchSummaryFeaturizer:
             file_name_or_path = [
                 os.path.join(self.path_to_jsons, f)
                 for f in os.listdir(self.path_to_jsons)
-                if not f.startswith("t")
-                and not f.startswith(".")
-                and not os.path.isdir(f)
+                if not f.startswith("t") and not f.startswith(".") and not os.path.isdir(f)
             ]
 
         elif self.path_to_lobster_calcs and not self.path_to_jsons:
@@ -361,15 +350,9 @@ class BatchSummaryFeaturizer:
         row = []
         with (
             mp.Pool(processes=self.n_jobs, maxtasksperchild=1) as pool,
-            tqdm(
-                total=len(file_name_or_path), desc="Generating LobsterPy summary stats"
-            ) as pbar,
+            tqdm(total=len(file_name_or_path), desc="Generating LobsterPy summary stats") as pbar,
         ):
-            for _, result in enumerate(
-                pool.imap_unordered(
-                    self._featurizelobsterpy, file_name_or_path, chunksize=1
-                )
-            ):
+            for _, result in enumerate(pool.imap_unordered(self._featurizelobsterpy, file_name_or_path, chunksize=1)):
                 pbar.update()
                 row.append(result)
 
@@ -387,13 +370,9 @@ class BatchSummaryFeaturizer:
         row = []
         with (
             mp.Pool(processes=self.n_jobs, maxtasksperchild=1) as pool,
-            tqdm(
-                total=len(paths), desc="Generating COHP/COOP/COBI summary stats"
-            ) as pbar,
+            tqdm(total=len(paths), desc="Generating COHP/COOP/COBI summary stats") as pbar,
         ):
-            for _, result in enumerate(
-                pool.imap_unordered(self._featurizecoxx, paths, chunksize=1)
-            ):
+            for _, result in enumerate(pool.imap_unordered(self._featurizecoxx, paths, chunksize=1)):
                 pbar.update()
[…]
