+++ lobsterpy/cli.py
@@ -32,9 +32,7 @@ def main() -> None:
 
 def get_parser() -> argparse.ArgumentParser:
     """Construct argumentparser with subcommands and sections."""
-    parser = argparse.ArgumentParser(
-        description="Analyze and plot results from Lobster runs."
-    )
+    parser = argparse.ArgumentParser(description="Analyze and plot results from Lobster runs.")
 
     # Arguments that are needed by different actions, but not always
 
@@ -129,9 +127,7 @@ def get_parser() -> argparse.ArgumentParser:
     calc_quality_description_file_parent = argparse.ArgumentParser(add_help=False)
 
     # Input args for specifically needed for calculation quality description
-    calc_quality_description_file_group = (
-        calc_quality_description_file_parent.add_argument_group("")
-    )
+    calc_quality_description_file_group = calc_quality_description_file_parent.add_argument_group("")
     calc_quality_description_file_group.add_argument(
         "-fvasprun",
         "--file-vasprun",
@@ -380,17 +376,14 @@ def get_parser() -> argparse.ArgumentParser:
     # Arguments specific to lobsterpy.cohp.analyze.Analysis and
     # lobsterpy.cohp.describe.Description class
     auto_parent = argparse.ArgumentParser(add_help=False)
-    auto_group = auto_parent.add_argument_group(
-        "Adjustable automatic analysis parameters"
-    )
+    auto_group = auto_parent.add_argument_group("Adjustable automatic analysis parameters")
     auto_group.add_argument(
         "-allb",
         "--allbonds",
         "--all-bonds",
         action="store_true",
         default=False,
-        help="Consider all bonds during the automatic analysis,"
-        " not only cation-anion bonds (default) ",
+        help="Consider all bonds during the automatic analysis, not only cation-anion bonds (default) ",
     )
     auto_group.add_argument(
         "--cutofficohp",
@@ -413,8 +406,7 @@ def get_parser() -> argparse.ArgumentParser:
         "--noisecutoff",
         type=float,
         default=None,
-        help="Sets the lower limit of icohps or icoops or icobis considered in"
-        " automatic analysis",
+        help="Sets the lower limit of icohps or icoops or icobis considered in automatic analysis",
     )
     auto_group.add_argument(
         "-orbresol",
@@ -457,9 +449,7 @@ def get_parser() -> argparse.ArgumentParser:
     )
     # Specific to interactive plotter args
     interactive_plotter_args = argparse.ArgumentParser(add_help=False)
-    interactive_plotter_group = interactive_plotter_args.add_argument_group(
-        "Options specific to interactive plotter"
-    )
+    interactive_plotter_group = interactive_plotter_args.add_argument_group("Options specific to interactive plotter")
     interactive_plotter_group.add_argument(
         "-labresol",
         "--labelresolved",
@@ -551,10 +541,7 @@ def get_parser() -> argparse.ArgumentParser:
             auto_parent,
             analysis_switch,
         ],
-        help=(
-            "Deliver a text description from automatic analysis of COHPs or COBIS or COOP results "
-            "from Lobster run"
-        ),
+        help=("Deliver a text description from automatic analysis of COHPs or COBIS or COOP results from Lobster run"),
     )
     subparsers.add_parser(
         "description-quality",
@@ -617,9 +604,7 @@ def get_parser() -> argparse.ArgumentParser:
             interactive_plotter_args,
             analysis_switch,
         ],
-        help=(
-            "Creates an interactive plot of most important COHPs or COBIs or COOPs automatically."
-        ),
+        help=("Creates an interactive plot of most important COHPs or COBIs or COOPs automatically."),
     )
     subparsers.add_parser(
         "plot-dos",
[…]
