+++ lobsterpy/featurize/core.py
@@ -63,9 +63,7 @@ class FeaturizeLobsterpy:
         """
         if self.path_to_json and not self.path_to_lobster_calc:
             # read the lightweight lobster json files using read_lobster_lightweight_json method
-            data = FeaturizeLobsterpy.read_lobster_lightweight_json(
-                path_to_json=self.path_to_json
-            )
+            data = FeaturizeLobsterpy.read_lobster_lightweight_json(path_to_json=self.path_to_json)
             if not ids:
                 ids = Path(self.path_to_json).name.split(".")[0]
 
@@ -81,9 +79,7 @@ class FeaturizeLobsterpy:
                 ids = Path(self.path_to_lobster_calc).name
 
         else:
-            raise ValueError(
-                "Please provide either path to lightweight lobster jsons or path to lobster calc"
-            )
+            raise ValueError("Please provide either path to lightweight lobster jsons or path to lobster calc")
         # define a pandas dataframe
         df = pd.DataFrame(index=[ids])
 
@@ -115,10 +111,7 @@ class FeaturizeLobsterpy:
             or not data[bond_type]["lobsterpy_data"]
             or not data[bond_type]["lobsterpy_data"]["sites"]
         ):
-            raise Exception(
-                f"No {self.bonds} bonds detected for {ids} structure. "
-                "Please switch to `all` bonds mode"
-            )
+            raise Exception(f"No {self.bonds} bonds detected for {ids} structure. Please switch to `all` bonds mode")
 
         if data[bond_type]["lobsterpy_data"]:
             for site_data in data[bond_type]["lobsterpy_data"]["sites"].values():
@@ -131,61 +124,35 @@ class FeaturizeLobsterpy:
                         if self.orbital_resolved:
                             if (
                                 bond_data["orbital_data"]["orbital_summary_stats"]
-                                and "max_bonding_contribution"
-                                in bond_data["orbital_data"]["orbital_summary_stats"]
+                                and "max_bonding_contribution" in bond_data["orbital_data"]["orbital_summary_stats"]
                             ):
-                                for orb_pair in bond_data["orbital_data"][
-                                    "orbital_summary_stats"
-                                ]["max_bonding_contribution"]:
-                                    icohp_mean_orb_bndg.append(
-                                        bond_data["orbital_data"][orb_pair][
-                                            "ICOHP_mean"
-                                        ]
-                                    )
-                                    icohp_sum_orb_bndg.append(
-                                        bond_data["orbital_data"][orb_pair]["ICOHP_sum"]
-                                    )
+                                for orb_pair in bond_data["orbital_data"]["orbital_summary_stats"][
+                                    "max_bonding_contribution"
+                                ]:
+                                    icohp_mean_orb_bndg.append(bond_data["orbital_data"][orb_pair]["ICOHP_mean"])
+                                    icohp_sum_orb_bndg.append(bond_data["orbital_data"][orb_pair]["ICOHP_sum"])
                                     bond_orb.append(
-                                        bond_data["orbital_data"][orb_pair][
[…]

[…]
