### 2024-02-23 670bf72
remove dead part of python-package.yml that is handled by pre-commit
