### 2024-02-24 JaGeo/LobsterPy#236 @naik-aakash
Improve help doc-string for create-inputs cli
Part of review comments of https://github.com/openjournals/joss-reviews/issues/6286#issuecomment-1961996406
