### 2024-02-24 JaGeo/LobsterPy#237 @JaGeo
In any case, thank you, @berquist ! I will close this then.
