### 2024-02-26 bba9e85
Update pymatgen requirement
Updates the requirements on [pymatgen](https://github.com/materialsproject/pymatgen) to permit the latest version.
- [Release notes](https://github.com/materialsproject/pymatgen/releases)
- [Changelog](https://github.com/materialsproject/pymatgen/blob/master/docs/CHANGES.md)
- [Commits](https://github.com/materialsproject/pymatgen/compare/v2023.9.10...v2024.2.23)

---
updated-dependencies:
- dependency-name: pymatgen
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
