### 2024-02-26 ca7e1d2
Rebuild file path handling in featurize module.
