+++ lobsterpy/featurize/batch.py
@@ -12,7 +12,6 @@ from pathlib import Path
 
 import numpy as np
 import pandas as pd
-from monty.os.path import zpath
 from tqdm.autonotebook import tqdm
 
 from lobsterpy.featurize.core import (
@@ -23,6 +22,8 @@ from lobsterpy.featurize.core import (
     FeaturizeLobsterpy,
 )
 from lobsterpy.structuregraph.graph import LobsterGraph
+
+from . import get_file_paths
 
 warnings.filterwarnings("ignore")
 
@@ -125,118 +126,57 @@ class BatchSummaryFeaturizer:
             Effective interaction number and moment features (center, width, skewness and kurtosis)
 
         """
-        dir_name = Path(path_to_lobster_calc)
+        file_paths = get_file_paths(
+            path_to_lobster_calc=path_to_lobster_calc, requested_files=["poscar", "cohpcar", "icohplist"]
+        )
+        structure_path = file_paths.get("poscar")
 
-        req_files = {
-            "structure_path": "POSCAR",
-            "coxxcar_path": "COHPCAR.lobster",
-            "icoxxlist_path": "ICOHPLIST.lobster",
-        }
-        for file, default_value in req_files.items():
-            # Check if "POSCAR" exists, and if not, check for "POSCAR.lobster"
-            if file == "structure_path":
-                for filename in [default_value, "POSCAR.lobster"]:
-                    poscar_path = dir_name / filename
-                    req_files[file] = poscar_path  # type: ignore
-                    if not poscar_path.exists():
-                        gz_file_path = Path(zpath(poscar_path))
-                        if gz_file_path.exists():
-                            req_files[file] = gz_file_path  # type: ignore
-                            break
-            else:
-                file_path = dir_name / default_value
-                req_files[file] = file_path  # type: ignore
-                if not file_path.exists():
-                    gz_file_path = Path(zpath(file_path))
-                    if gz_file_path.exists():
-                        req_files[file] = gz_file_path  # type: ignore
+        coxx = FeaturizeCOXX(
+            path_to_coxxcar=str(file_paths.get("cohpcar")),
+            path_to_icoxxlist=str(file_paths.get("icohp_list")),
+            path_to_structure=str(structure_path),
+            feature_type=self.feature_type,
+            e_range=self.e_range,
+        )
 
-        coxxcar_path = req_files.get("coxxcar_path")
-        structure_path = req_files.get("structure_path")
-        icoxxlist_path = req_files.get("icoxxlist_path")
+        df_cohp = coxx.get_summarized_coxx_df()
+        del coxx
 
-        if (
-            coxxcar_path.exists()  # type: ignore
-            and structure_path.exists()  # type: ignore
-            and icoxxlist_path.exists()  # type: ignore
-        ):
+        if self.include_cobi_data:
+            file_paths = get_file_paths(
+                path_to_lobster_calc=path_to_lobster_calc, requested_files=["cobicar", "icobilist"]
+            )
+
             coxx = FeaturizeCOXX(
-                path_to_coxxcar=str(coxxcar_path),
-                path_to_icoxxlist=str(icoxxlist_path),
+                path_to_coxxcar=str(file_paths.get("cobicar")),
+                path_to_icoxxlist=str(file_paths.get("icobilist")),
                 path_to_structure=str(structure_path),
                 feature_type=self.feature_type,
                 e_range=self.e_range,
+                are_cobis=True,
             )
 
-            df_cohp = coxx.get_summarized_coxx_df()
+            df_cobi = coxx.get_summarized_coxx_df()
             del coxx
-        else:
[…]
