### 2024-02-26 8c310f2
adapt calc_quality_summary to work with pymatgen objs
