+++ lobsterpy/cohp/analyze.py
@@ -13,6 +13,7 @@ from pymatgen.analysis.bond_valence import BVAnalyzer
 from pymatgen.core.structure import Structure
 from pymatgen.electronic_structure.cohp import CompleteCohp
 from pymatgen.electronic_structure.core import Spin
+from pymatgen.electronic_structure.dos import LobsterCompleteDos
 from pymatgen.io.lobster import (
     Bandoverlaps,
     Charge,
@@ -1482,15 +1483,22 @@ class Analysis:
 
     @staticmethod
     def get_lobster_calc_quality_summary(
-        path_to_poscar: str,
-        path_to_lobsterout: str,
-        path_to_lobsterin: str,
+        path_to_poscar: str | None = None,
+        path_to_lobsterout: str | None = None,
+        path_to_lobsterin: str | None = None,
         path_to_potcar: str | None = None,
         potcar_symbols: list | None = None,
         path_to_charge: str | None = None,
         path_to_bandoverlaps: str | None = None,
         path_to_doscar: str | None = None,
         path_to_vasprun: str | None = None,
+        structure_obj: Structure | None = None,
+        lobsterin_obj: Lobsterin | None = None,
+        lobsterout_obj: Lobsterout | None = None,
+        charge_obj: Charge | None = None,
+        bandoverlaps_obj: Bandoverlaps | None = None,
+        lobster_completedos_obj: LobsterCompleteDos | None = None,
+        vasprun_obj: Vasprun | None = None,
         dos_comparison: bool = False,
         e_range: list = [-5, 0],
         n_bins: int | None = None,
@@ -1508,6 +1516,13 @@ class Analysis:
         :param path_to_bandoverlaps: path to bandOverlaps.lobster file
         :param path_to_doscar: path to DOSCAR.lobster or DOSCAR.LSO.lobster file
         :param path_to_vasprun: path to vasprun.xml file
+        :param structure_obj: pymatgen pymatgen.core.structure.Structure object
+        :param lobsterin_obj: pymatgen.lobster.io.Lobsterin object
+        :param lobsterout_obj: pymatgen lobster.io.Lobsterout object
+        :param charge_obj: pymatgen lobster.io.Charge object
+        :param bandoverlaps_obj: pymatgen lobster.io.BandOverlaps object
+        :param lobster_completedos_obj: pymatgen.electronic_structure.dos.LobsterCompleteDos object
+        :param vasprun_obj: pymatgen vasp.io.Vasprun object
         :param dos_comparison: will compare DOS from VASP and LOBSTER and return tanimoto index
         :param e_range: energy range for DOS comparisons
         :param n_bins: number of bins to discretize DOS for comparisons
@@ -1521,23 +1536,54 @@ class Analysis:
         """
         quality_dict = {}
 
-        if path_to_potcar and not potcar_symbols:
+        if (
+            path_to_potcar
+            and not potcar_symbols
+            and not path_to_vasprun
+            and not vasprun_obj
+        ):
             potcar_names = Lobsterin._get_potcar_symbols(POTCAR_input=path_to_potcar)
-        elif not path_to_potcar and potcar_symbols:
+        elif (
+            not path_to_potcar
+            and not path_to_vasprun
+            and not vasprun_obj
+            and potcar_symbols
+        ):
             potcar_names = potcar_symbols
+        elif path_to_vasprun and not vasprun_obj:
+            vasprun = Vasprun(
+                path_to_vasprun, parse_potcar_file=False, parse_eigen=False
+            )
+            potcar_names = [potcar.split(" ")[1] for potcar in vasprun.potcar_symbols]
+        elif vasprun_obj and not path_to_vasprun:
+            potcar_names = [
[…]
