+++ tests/cohp/test_analyze.py
@@ -835,10 +835,7 @@ class TestAnalyse:
             analyse_nacl_comp_range_orb.condensed_bonding_analysis
             == analyse_nacl_comp_range_orb_with_objs.condensed_bonding_analysis
         )
-        assert (
-            analyse_k3sb_all.condensed_bonding_analysis
-            == analyse_k3sb_all_objs.condensed_bonding_analysis
-        )
+        assert analyse_k3sb_all.condensed_bonding_analysis == analyse_k3sb_all_objs.condensed_bonding_analysis
 
     def test_icohp_sum_nacl(self, analyse_nacl_comp_range):
         assert abs(
