### 2024-02-26 42500d7
Added charge_type checking in BatchSummaryFeaturizer, changed missing structure file handling in featurizer module.
