### 2024-02-26 b7dfa0b
Cleaned up _featurizecoxx() of BatchSummaryFeaturizer, corrected charge_type handling in BatchSummaryFeaturizer.
