### 2024-02-26 1561f97
Added parameter checks for all str type parameters of BatchSummaryFeaturizer.
