### 2024-02-26 4a5d974
Corrected possible bond type parameters.
