### 2024-02-26 9514f8a
Moved file path handling to separate utils.py, added test for structure path handling.
