### 2024-02-26 c981db1
cover missing doscar exception
