+++ tests/cohp/test_analyze.py
@@ -993,8 +993,9 @@ class TestAnalyseCalcQuality:
                 structure_obj=structure_obj,
                 lobster_completedos_obj=None,
                 charge_obj=charge_obj,
+                potcar_symbols=["K_sv", "Sb"],
                 bandoverlaps_obj=bandoverlaps_obj,
-                vasprun_obj=vasprun_obj,
+                vasprun_obj=None,
                 lobsterin_obj=lobsterin_obj,
                 lobsterout_obj=lobsterout_obj,
                 bva_comp=True,
