### 2024-02-26 7d28e22
Modified Exception of featurizer utils, adapted featurizer exception tests.
