### 2024-02-26 2b71ee0
Fixed failing core exception test. Adapted charge_type error message to display actual possible parameter values.
