### 2024-02-26 96b833b
Added test for checking str parameters in init of BatchSummaryFeaturizer.
