### 2024-02-26 4c372dd
extend help in cli for create inputs
