### 2024-02-26 6e7e5af
add more explanation for the limitation of create-inputs
