+++ lobsterpy/cli.py
@@ -531,7 +531,7 @@ def get_parser() -> argparse.ArgumentParser:
         ],
         help="Create inputs for lobster computation. It works only with PBE POTCARs. "
         "As currently only pbeVASPfit2015 basis in LOBSTER includes additional orbitals relevant for "
-        "the solid state materials. Please check our publication https://doi.org/10.1002/cplu.202200123 "
+        "the solid state materials. Please check out our publication https://doi.org/10.1002/cplu.202200123 "
         "and LOBSTER program manual for more information.",
     )
     subparsers.add_parser(
