### 2024-02-26 279d5c7
Added further tests for file path handling in featurizer module.
