### 2024-02-26 7697f48
Corrected ruff SIM117.
