### 2024-02-26 4513bc9
add create-inputs examples
