### 2024-02-26 c2db763
Update lobsterpy/cli.py
Co-authored-by: J. George <JaGeo@users.noreply.github.com>
