+++ lobsterpy/cli.py
@@ -529,9 +529,9 @@ def get_parser() -> argparse.ArgumentParser:
             user_basis_arg,
             output_parent,
         ],
-        help="Create inputs for lobster computation. It works only with PBE POTCARs. "
-        "As currently only pbeVASPfit2015 basis in LOBSTER includes additional orbitals relevant for "
-        "the solid state materials. Please check out our publication https://doi.org/10.1002/cplu.202200123 "
+        help="Create inputs for lobster computation. It works only with PBE POTCARs, "
+        "as, currently, only the pbeVASPfit2015 basis fitted to PBE POTCARs in LOBSTER includes additional orbitals relevant for "
+        "the solid-state materials. Please check out our publication https://doi.org/10.1002/cplu.202200123 "
         "and LOBSTER program manual for more information.",
     )
     subparsers.add_parser(
