### 2024-02-26 0315c94
add cutoff args to batch structure graphs featurizer
