+++ lobsterpy/plotting/layout_dicts.py
@@ -32,6 +32,7 @@ layout_dict = {
     :param str paper_bgcolor: background color.
     :param str plot_bgcolor: plot background color.
 """
+
 cohp_axis_style_dict = {
     "titlefont": {"size": 20},
     "tickfont": {"size": 16},
@@ -62,6 +63,7 @@ cohp_axis_style_dict = {
     :param float linewidth: width of axis line.
     :param str linecolor: color of axis line.
 """
+
 energy_axis_style_dict = {
     "titlefont": {"size": 20},
     "tickfont": {"size": 16},
@@ -92,6 +94,7 @@ energy_axis_style_dict = {
     :param float linewidth: width of axis line.
     :param str linecolor: color of axis line.
 """
+
 spin_up_trace_style_dict = {
     "line": {"width": 3},
     "mode": "lines",
@@ -117,6 +120,7 @@ spin_down_trace_style_dict = {
     :param str mode: Plotly mode.
     :param str visible: Trace visibility.
 """
+
 legend_style_dict = {
     "bordercolor": "black",
     "borderwidth": 2,
