+++ lobsterpy/plotting/layout_dicts.py
@@ -32,7 +32,6 @@ layout_dict = {
     :param str paper_bgcolor: background color.
     :param str plot_bgcolor: plot background color.
 """
-
 cohp_axis_style_dict = {
     "titlefont": {"size": 20},
     "tickfont": {"size": 16},
@@ -63,7 +62,6 @@ cohp_axis_style_dict = {
     :param float linewidth: width of axis line.
     :param str linecolor: color of axis line.
 """
-
 energy_axis_style_dict = {
     "titlefont": {"size": 20},
     "tickfont": {"size": 16},
@@ -94,7 +92,6 @@ energy_axis_style_dict = {
     :param float linewidth: width of axis line.
     :param str linecolor: color of axis line.
 """
-
 spin_up_trace_style_dict = {
     "line": {"width": 3},
     "mode": "lines",
@@ -107,7 +104,6 @@ spin_up_trace_style_dict = {
     :param str mode: Plotly mode.
     :param str visible: Trace visibility.
 """
-
 spin_down_trace_style_dict = {
     "line": {"width": 3, "dash": "dash"},
     "mode": "lines",
@@ -120,7 +116,6 @@ spin_down_trace_style_dict = {
     :param str mode: Plotly mode.
     :param str visible: Trace visibility.
 """
-
 legend_style_dict = {
     "bordercolor": "black",
     "borderwidth": 2,
