### 2024-02-26 #6286 @RMeli
@JaGeo I see you have the repository already mirrored on Zenodo. Would it make sense to add the name of the software to the Zenodo title? Something like `LobsterPy: Automated Bonding Analysis with Crystal Orbital Hamilton Populations`?
