### 2024-02-26 #6286 @JaGeo
> @JaGeo Is affiliation (1) correct? "Department Materials Chemistry" sounds a bit odd to me (but I'm not a native English speaker). I see that on the [BAM website](https://www.bam.de/Navigation/EN/About-us/Organisation/Organisation-Chart/President/Department-6/department-6.html) it is called "Materials Chemistry department"? Or maybe "Department 6: Materials Chemistry" would be more appropriate? If it's correct as it is, please discard my comment.

We can surely replace it with Materials Chemistry Department! Thanks for the suggestion!
