+++ paper/paper.md
@@ -33,21 +33,21 @@ bibliography: paper.bib
 
 ---
 # Summary
-The LOBSTER[@deringer2011crystal;@maintz2013analytic;@maintz2016lobster;@nelson2020lobster] software aids in extracting quantum-chemical bonding information from materials by projecting the
+The LOBSTER [@deringer2011crystal;@maintz2013analytic;@maintz2016lobster;@nelson2020lobster] software aids in extracting quantum-chemical bonding information from materials by projecting the
 plane-wave based wave functions from density functional theory (DFT) onto an atomic orbital basis. [LobsterEnv](https://github.com/materialsproject/pymatgen/blob/master/pymatgen/io/lobster/lobsterenv.py),
-a module implemented in pymatgen[@ong2013python] by some of the authors of this package, facilitates the use of quantum-chemical bonding
+a module implemented in pymatgen [@ong2013python] by some of the authors of this package, facilitates the use of quantum-chemical bonding
 information obtained from LOBSTER calculations to identify neighbors and coordination environments. _LobsterPy_ is a Python package that offers a set of convenient tools
 to further analyze and summarize the LobsterEnv outputs in the form of JSONs that are easy to interpret and process. These tools enable the
 estimation of (anti) bonding contributions, generation of textual descriptions, and visualization of LOBSTER computation results.  Since its first release, both _LobsterPy_ and _LobsterEnv_ capabilities
-have been extended significantly. Unlike earlier versions, which could only automatically analyze Crystal Orbital Hamilton Populations (COHPs)[@dronskowski1993crystal],
-both can now also analyze Crystal Orbital Overlap Populations (COOP)[@hughbanks1983chains] and Crystal Orbital Bond Index (COBI)[@mueller2021crystal].
+have been extended significantly. Unlike earlier versions, which could only automatically analyze Crystal Orbital Hamilton Populations (COHPs) [@dronskowski1993crystal],
+both can now also analyze Crystal Orbital Overlap Populations (COOP) [@hughbanks1983chains] and Crystal Orbital Bond Index (COBI) [@mueller2021crystal].
 Extracting the information about the most important orbitals contributing to the bonds is optional, and users can enable it as needed.
 Additionally, bonding-based features for machine-learning (ML) studies can be engineered via the sub-packages "featurize" and "structuregraphs".
 Alongside its Python interface, it also provides an easy-to-use command line interface (CLI) that runs automatic analysis of the
 computations and generates a summary of results and publication-ready figures.
 
-_LobsterPy_ has been used to produce the results in [@ngo2023dft; @chen2024insights; @naik2023quantumchemical] and is also part of
-[@atomate2] bonding analysis workflow for generating bonding analysis data in a format compatible with the Materials Project[@materialsproject] API.
+_LobsterPy_ has been used to produce the results in @ngo2023dft, @chen2024insights, @naik2023quantumchemical, and it is also part of
+@atomate2 bonding analysis workflow for generating bonding analysis data in a format compatible with the Materials Project [@materialsproject] API.
 
 # Statement of need
 Although the notion of "bonds" might seem unusual from a physicist's point of view, chemists have been employing it routinely to
@@ -69,7 +69,7 @@ _LobsterPy_ aims to fulfill this need.
 - Perform automatic bonding analysis and plotting via inherent command line interface app.
 
 # Similar and Related Software
[…]
