### 2024-02-26 #6286 @JaGeo
@RMeli :

Zenodo:
https://zenodo.org/records/10631237
Would this be fine? I would repeat this change for our next release as well.

Affiliation:
I changed the affiliation as suggested: https://github.com/JaGeo/LobsterPy/blob/main/paper/paper.md
