### 2024-02-26 #6286 @RMeli
> Zenodo:
https://zenodo.org/records/10631237
Would this be fine? I would repeat this change for our next release as well.

Looks great, I think this will be much easier to find.
