### 2024-02-26 #6286 @JaGeo
> @JaGeo I see that the author list on Zenodo is also different from the author list of the paper. The JOSS editorial guide states that
>
> > Check the archive deposit has the correct metadata (title and author list), and request the author edit it if it doesn’t match the paper.
>
> This is my first time editing, I'll ask internally if this is an hard rule but as a minimum the authors in the paper should be all included in the Zenodo citation (I think); the other way around I don't think should hold. Will the missing authors be added with the next release or do you need to change this manually?

I will change this manually. The problem comes from the `CITATION.cff` file in the repo. Zenodo uses it as a basis for the citation. I am happy to do this correctly for the next release!
