### 2024-02-26 #6286 @RMeli
I asked internally how it works. I hope that the Zenodo list can be a superset (i.e. all contributors) of the JOSS authors (main/core contributors). I'll let you know ASAP how to proceed.
