### 2024-02-26 afd084d
Update paper/paper.md
