### 2024-02-26 #6286 @JaGeo
@RMeli As a part of this PR that will also fix some tiny issues with the README and our code coverage determination, we have added that acknowledgment now: https://github.com/JaGeo/LobsterPy/pull/246
