### 2024-02-26 d0f8ce4
Fix spelling error in  CHANGELOG.md
