### 2024-02-27 #6286 @Kevin-Mattheus-Moerman
@RMeli thanks for editing this one. It looks like perhaps `@editorialbot create post-review checklist` was not used yet. I'll call it now and take it from here. I recommend using it in the future to prevent overlooking some of the last steps. In this case for instance the license on the archive link doesn't match up yet. No worries though we'll sort this out now. Thanks again.
