### 2024-02-27 #6286 @Kevin-Mattheus-Moerman
@JaGeo As AEiC on this track I will now help process the final steps to help with acceptance in JOSS. I have just checked the paper, this review, your repository, and the archive link. Most seems in order. I'd only ask you to please ensure that the license listed on the archive link matches your software license. Can you edit this please?
