### 2024-03-06 1da0a8a
remove redundant orbplot arg and fix a typo
